// packages/db/migrations/snapshots/27110bc5899087831e3bf607bc6e68f1ddb630ea69fee986e3f38cffb6bbf475/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Member: {
      readonly avatarKey: CodecTypes['pg/text@1']['output'] | null;
      readonly avatarUrl: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly role: 'owner';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly workspaceId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly Workspace: {
      readonly agentAvatarKey: CodecTypes['pg/text@1']['output'] | null;
      readonly agentEnabled: CodecTypes['pg/bool@1']['output'];
      readonly agentName: CodecTypes['pg/text@1']['output'] | null;
      readonly allowedDomains: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly color: 'blue' | 'indigo' | 'violet' | 'pink' | 'red' | 'orange' | 'green';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly greeting: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly position: 'bottom-left' | 'bottom-right';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Member: {
      readonly avatarKey: CodecTypes['pg/text@1']['input'] | null;
      readonly avatarUrl: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly role: 'owner';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly workspaceId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly Workspace: {
      readonly agentAvatarKey: CodecTypes['pg/text@1']['input'] | null;
      readonly agentEnabled: CodecTypes['pg/bool@1']['input'];
      readonly agentName: CodecTypes['pg/text@1']['input'] | null;
      readonly allowedDomains: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly color: 'blue' | 'indigo' | 'violet' | 'pink' | 'red' | 'orange' | 'green';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly greeting: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly position: 'bottom-left' | 'bottom-right';
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly members: {
      readonly avatar_key: CodecTypes['pg/text@1']['output'] | null;
      readonly avatar_url: CodecTypes['pg/text@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly role: 'owner';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly user_id: CodecTypes['pg/uuid@1']['output'];
      readonly workspace_id: CodecTypes['pg/uuid@1']['output'];
    };
    readonly workspaces: {
      readonly agent_avatar_key: CodecTypes['pg/text@1']['output'] | null;
      readonly agent_enabled: CodecTypes['pg/bool@1']['output'];
      readonly agent_name: CodecTypes['pg/text@1']['output'] | null;
      readonly allowed_domains: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly color: 'blue' | 'indigo' | 'violet' | 'pink' | 'red' | 'orange' | 'green';
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly greeting: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly position: 'bottom-left' | 'bottom-right';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly members: {
      readonly avatar_key: CodecTypes['pg/text@1']['input'] | null;
      readonly avatar_url: CodecTypes['pg/text@1']['input'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly role: 'owner';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly user_id: CodecTypes['pg/uuid@1']['input'];
      readonly workspace_id: CodecTypes['pg/uuid@1']['input'];
    };
    readonly workspaces: {
      readonly agent_avatar_key: CodecTypes['pg/text@1']['input'] | null;
      readonly agent_enabled: CodecTypes['pg/bool@1']['input'];
      readonly agent_name: CodecTypes['pg/text@1']['input'] | null;
      readonly allowed_domains: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly color: 'blue' | 'indigo' | 'violet' | 'pink' | 'red' | 'orange' | 'green';
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly greeting: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly position: 'bottom-left' | 'bottom-right';
      readonly updated_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;